requirements/base.txt
aiodns==3.0.0
    # via -r base.in
aiohttp==3.8.4
    # via
    #   -r base.in
    #   aiohttp-json-rpc
    #   discord-py
    #   red-lavalink
aiohttp-json-rpc==0.13.3
    # via -r base.in
aiosignal==1.3.1
    # via aiohttp
apsw==3.41.2.0
    # via -r base.in
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
babel==2.12.1
    # via -r base.in
brotli==1.0.9
    # via -r base.in
cffi==1.15.1
    # via pycares
charset-normalizer==3.1.0
    # via aiohttp
click==8.1.3
    # via -r base.in
contextlib2==21.6.0
    # via schema
discord-py==2.2.3
    # via
    #   -r base.in
    #   red-lavalink
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
importlib-metadata==6.6.0
    # via markdown
markdown==3.4.3
    # via -r base.in
markdown-it-py==2.2.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
orjson==3.8.12
    # via -r base.in
packaging==23.1
    # via -r base.in
platformdirs==3.5.1
    # via -r base.in
psutil==5.9.5
    # via -r base.in
pycares==4.3.0
    # via aiodns
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via rich
python-dateutil==2.8.2
    # via -r base.in
pytz==2023.3
    # via babel
pyyaml==6.0
    # via -r base.in
rapidfuzz==3.0.0
    # via -r base.in
red-commons==1.0.0
    # via
    #   -r base.in
    #   red-lavalink
red-lavalink==0.11.0
    # via -r base.in
rich==13.3.5
    # via -r base.in
schema==0.7.5
    # via -r base.in
six==1.16.0
    # via python-dateutil
tabulate==0.9.0
    # via -r base.in
typing-extensions==4.5.0
    # via
    #   -r base.in
    #   rich
yarl==1.9.2
    # via
    #   -r base.in
    #   aiohttp
zipp==3.15.0
    # via importlib-metadata
colorama==0.4.6; sys_platform == "win32"
    # via click
distro==1.8.0; sys_platform == "linux" and sys_platform == "linux"
    # via -r base.in
uvloop==0.17.0; (sys_platform != "win32" and platform_python_implementation == "CPython") and sys_platform != "win32"
    # via -r base.in
